Fixed story tile and text heading Fixed button page
